+++ src/linalg-cpu.h
@@ -3,9 +3,9 @@
 
 #include "lib.h"
 
-void vector_dot_product_cpu(Tensor* tensorA, Tensor* tensorB, double* result_data);
-void matmul_2d_2d_cpu(Tensor* tensorA, Tensor* tensorB, double* result_data);
-void matmul_broadcasted_cpu(Tensor* tensorA, Tensor* tensorB, double* result_data);
-void outer_cpu(Tensor* tensorA, Tensor* tensorB, double* result_data);
+void vector_dot_product_cpu(Tensor* tensorA, Tensor* tensorB, float* result_data);
+void matmul_2d_2d_cpu(Tensor* tensorA, Tensor* tensorB, float* result_data);
+void matmul_broadcasted_cpu(Tensor* tensorA, Tensor* tensorB, float* result_data);
+void outer_cpu(Tensor* tensorA, Tensor* tensorB, float* result_data);
 
 #endif
